Move libp2p config into separate module
