+++ aquadoggo/src/libp2p/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod config;
 mod service;
 
 pub use service::libp2p_service;
